Added all relevant files and python script
The main script that solves the problem is amadeus_task_oscar_nieves.py. There are 2 pdf files which explain in detail how the problem is solved, how the algorithm works and how the Python script should be used. The text files contain the separate test cases for ease of use.